registration back

+++ src/main/java/com/epam/webappfinal/command/DeleteCommentItemCommand.java
@@ -1,15 +1,18 @@
 package com.epam.webappfinal.command;
 
+import com.epam.webappfinal.entity.User;
 import com.epam.webappfinal.exception.ServiceException;
 import com.epam.webappfinal.service.FoodService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 public class DeleteCommentItemCommand implements Command {
 
     private static final String COMMENT_ID_TEXT_REPRESENTATION = "commentId";
     private static final String RATING_PAGE_COMMAND = "controller?command=rating";
+    private static final String LOGIN_PAGE = "/login.jsp";
 
     private final FoodService foodService;
 
@@ -19,9 +22,16 @@ public class DeleteCommentItemCommand implements Command {
 
     @Override
     public CommandResult execute(HttpServletRequest req, HttpServletResponse resp) throws ServiceException {
-        String commentIdStr = req.getParameter(COMMENT_ID_TEXT_REPRESENTATION);
-        Long commentId = Long.parseLong(commentIdStr);
-        foodService.deleteComment(commentId);
-        return CommandResult.redirect(RATING_PAGE_COMMAND);
+        HttpSession session = req.getSession();
+        User user = (User) session.getAttribute(User.TABLE_NAME);
+        if (user != null) {
+            String commentIdStr = req.getParameter(COMMENT_ID_TEXT_REPRESENTATION);
+            Long commentId = Long.parseLong(commentIdStr);
+            foodService.deleteComment(commentId);
+            return CommandResult.redirect(RATING_PAGE_COMMAND);
+        } else {
+            return CommandResult.forward(LOGIN_PAGE);
+        }
+
     }
 }
